search/pyproject.toml
[project]
name = "indexer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "insightface>=0.7.3",
    "onnxruntime>=1.22.1",
    "pydantic-settings>=2.10.1",
    "pydantic-settings-yaml>=0.2.0",
    "pymilvus>=2.6.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
